+++ packages/dashboard/docs/stories/TilePicker.story.js
@@ -1,11 +1,17 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
 import { Application } from 'react-rainbow-components';
+import { FilterFilled } from '@rainbow-modules/icons';
 import { TilePicker, Tile } from '../../src';
 
 const StyledContainer = styled.div`
     padding: 25px 10px;
     background-color: #f4f6f9;
+`;
+
+const StyledFilterFilled = styled(FilterFilled)`
+    width: 8px;
+    height: 8px;
 `;
 
 export const OnlyOneSelection = () => {
@@ -95,6 +101,98 @@ export const MultipleSelection = () => {
                         value={8}
                         color="#ffffff"
                         backgroundColor="#f2707a"
+                    />
+                </TilePicker>
+            </StyledContainer>
+        </Application>
+    );
+};
+
+export const BadgeVariant = () => {
+    const [value, setValue] = useState();
+    return (
+        <Application>
+            <StyledContainer>
+                <TilePicker value={value} onChange={setValue}>
+                    <Tile
+                        name="completed"
+                        label="Completed"
+                        value={12}
+                        color="#ffffff"
+                        backgroundColor="#9dc35e"
+                        variant="badge"
+                    />
+                    <Tile
+                        name="assigned"
+                        label="Assigned"
+                        value={32}
+                        color="#ffffff"
+                        backgroundColor="#71d4d6"
+                        variant="badge"
+                    />
+                    <Tile
+                        name="pending"
+                        label="Pending"
+                        value={23}
+                        color="#ffffff"
+                        backgroundColor="#f4d87a"
+                        variant="badge"
+                    />
+                    <Tile
+                        name="canceled"
+                        label="Canceled"
+                        value={8}
+                        color="#ffffff"
+                        backgroundColor="#f2707a"
+                        variant="badge"
+                    />
+                </TilePicker>
+            </StyledContainer>
+        </Application>
+    );
+};
+
+export const BadgeVariantWithCustomIcon = () => {
+    const [value, setValue] = useState();
+    return (
+        <Application>
+            <StyledContainer>
+                <TilePicker value={value} onChange={setValue}>
+                    <Tile
+                        name="completed"
+                        label="Completed"
+                        value={12}
+                        color="#ffffff"
+                        backgroundColor="#9dc35e"
+                        variant="badge"
+                        selectedIcon={<StyledFilterFilled />}
+                    />
+                    <Tile
+                        name="assigned"
+                        label="Assigned"
+                        value={32}
+                        color="#ffffff"
+                        backgroundColor="#71d4d6"
+                        variant="badge"
+                        selectedIcon={<StyledFilterFilled />}
+                    />
+                    <Tile
+                        name="pending"
+                        label="Pending"
+                        value={23}
+                        color="#ffffff"
+                        backgroundColor="#f4d87a"
+                        variant="badge"
+                        selectedIcon={<StyledFilterFilled />}
+                    />
+                    <Tile
+                        name="canceled"
+                        label="Canceled"
+                        value={8}
+                        color="#ffffff"
+                        backgroundColor="#f2707a"
+                        variant="badge"
+                        selectedIcon={<StyledFilterFilled />}
                     />
                 </TilePicker>
             </StyledContainer>