+++ packages/dashboard/src/components/Tile/__test__/tile.spec.js
@@ -3,7 +3,7 @@ import { mount } from 'enzyme';
 import { Application } from 'react-rainbow-components';
 import Tile from '..';
 import { Provider } from '../../TilePicker/context';
-import { StyledContent, StyledValue } from '../styled';
+import { StyledContent } from '../styled';
 
 describe('<Tile />', () => {
     it('should render a simple div content when do not have TilePicker context', () => {
@@ -67,27 +67,5 @@ describe('<Tile />', () => {
             .first()
             .simulate('change', { target: { checked: true } });
         expect(onChangeMock).toBeCalledWith(name, true);
-    });
-
-    it('should render a component to the background color passed ', () => {
-        const backgroundColor = '#000000';
-        const wrapper = mount(
-            <Application>
-                <Tile backgroundColor={backgroundColor} />
-            </Application>,
-        );
-        expect(wrapper.find(StyledContent).first().prop('style')).toStrictEqual({
-            backgroundColor,
-        });
-    });
-
-    it('should render a component to the color passed', () => {
-        const color = '#ffffff';
-        const wrapper = mount(
-            <Application>
-                <Tile color={color} />
-            </Application>,
-        );
-        expect(wrapper.find(StyledValue).first().prop('style')).toStrictEqual({ color });
     });
 });