+++ packages/forms/docs/stories/UniversalPicker.story.js
@@ -15,6 +15,7 @@ const Container = styled.div`
     display: flex;
     flex-direction: column;
     justify-content: center;
+    align-items: center;
 `;
 
 const StyledOption = attachThemeAttrs(styled.span)`