+++ packages/listview/src/index.js
@@ -2,3 +2,4 @@
 export { default as FirestoreTable } from './components/FirestoreTable';
 export { default as ColoredStatusColumn } from './components/ColoredStatusColumn';
 export { default as BatchActionsBar } from './components/BatchActionsBar';
+export { default as NavigationButtonColumn } from './components/NavigationButtonColumn';