+++ packages/forms/CHANGELOG.md
@@ -3,6 +3,21 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.8.0](https://github.com/nexxtway/rainbow-modules/compare/v0.7.1...v0.8.0) (2020-10-27)
+
+**Note:** Version bump only for package @rainbow-modules/forms
+
+# [0.7.0](https://github.com/nexxtway/rainbow-modules/compare/v0.6.0...v0.7.0) (2020-10-26)
+
+**Note:** Version bump only for package @rainbow-modules/forms
+
+# [0.6.0](https://github.com/nexxtway/rainbow-modules/compare/v0.4.0...v0.6.0) (2020-10-26)
+
+### Features
+
+-   add labelAlignment prop on the UniversalPicker component ([#171](https://github.com/nexxtway/rainbow-modules/issues/171)) ([ba5e521](https://github.com/nexxtway/rainbow-modules/commit/ba5e521e4c4a2f9c107fc37bccac830f3c48ae5d)), closes [#165](https://github.com/nexxtway/rainbow-modules/issues/165)
+-   implement StripeCheckoutModal ([#170](https://github.com/nexxtway/rainbow-modules/issues/170)) ([acb66e7](https://github.com/nexxtway/rainbow-modules/commit/acb66e7965d642c1d5228eabfeffe67d4656a931))
+
 # [0.5.0](https://github.com/nexxtway/rainbow-modules/compare/v0.4.0...v0.5.0) (2020-10-22)
 
 **Note:** Version bump only for package @rainbow-modules/forms